GamePage component separated
